Create PowerBI
My last project on the theme of the Junior School Data.
Using Power BI I created several different charts and created a basic DAX measure to use for the Card Visual.  
On a couple of the charts I used conditional formatting to create uniformity as I thought it warranted it.